- Refactor

+++ src/main/java/com/khryniewicki/projectX/game/attack/spells/spell_properties/Spell.java
@@ -42,8 +42,8 @@ public class Spell implements UltraSpell {
     public float[] tcs;
     private float indexHeight = 1;
     private float indexWidth = 1;
-    private float powerAttack;
-    private float manaConsumed;
+    private Integer powerAttack;
+    private Integer manaConsumed;
 
     private Application.MyStompSessionHandler application = new Application.MyStompSessionHandler();
